351001/Budnikov/publisher/pyproject.toml
[project]
name = "budnikov"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.3"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "tortoise-orm>=1.1.0",
    "uvicorn>=0.44.0",
]

[tool.ruff]
line-length = 120

[tool.uv.workspace]
members = [
    "discussion",
]

